manually use `aapt singleCrunch` for reproducible builds
PNG crunching is not a deterministic process, especially the way aapt does
it.  This makes the builds not reproducible.  The easy solution to this is
to pre-crunch the PNGs and commit them to git.  This also uses exiftool to
strip any metadata from the PNG.  This then disables running the crunch as
part of the gradle build process.

https://medium.com/@duhroach/smaller-pngs-and-android-s-aapt-tool-4ce38a24019d