strip trailing whitespace from TorService
I use Android Studio with the defaults, and it seems to be stripping the
trailing whitespace off, which makes it annoying to see what has changed.
So this just commits the removal of trailing whitespace.

If you want to do it for the whole project, you can use this:
find * -name \*.java -print0 |xargs -0 sed -i 's,\s*$,,'