Fixes #523 Onion Site Backup Bags
This bug was introdued when a previous bug fix pertaining to Onion service storage in Orbot was addressed.
The old bug made it such that multiple onion services could be hosted that pointed to the same local port on the device.
This previous bug fix broke the backup logic for v2 and v3 onion services, which is fixed here.