Start Orbot via always-on VPN
When the system tries to start Orbot as a result of the always-on VPN
setting, automatically connect to Tor as a VPN without requiring
additional user intervention. In this case, the system will also start
Orbot on boot regardless of the "Start Orbot on Boot" setting.
This support is especially important for scenarios where Orbot is
bundled with the OS in always-on mode.

Test: Manual: Install Orbot and start the VPN. In the system settings,
set Orbot as an always-on VPN. Clear storage for Orbot, and optionally
turn on Notifiations. Reboot the device, or turn the always-on VPN
setting off and on again. In either case, Orbot automatically connects
and is available for use.