For now disable app rotation.
There are many ways the app crashes when the device is rotated. There are many layouts, especially
on phones, that are unusable/need a ton of work when in landscape.

This commit locks orbot into portrait modes on phone.
On tablets, orbot will be locked in portrait or landscape (but you can't rotate the app to change
the orientation still)

Open to contributors to get this functionality back :)